[IMP] outlook: correctly show a tooltip if the email is truncated
Purpose
=======
Currently, the email address of the opened email can be truncated if
it's too big. If the email is truncated, a tooltip can be visible when
we hover the email. It was done in JS with a code that compute the size
in pixel of the element... Which is very ugly.

Hopefully, an option on the component "TooltipHost" exists
just for that and can be used instead.

Task-2601837