[REF] outlook: clean the company widget and the partner widget
Purpose
=======
This commit clean the Company and the Partner widget.

Company Widget
==============
The "company name" widget inherited from the "partner widget" which
add a lot of complexity for nothing. Some code specific to the company
were in the "partner widget" (like `isParentCompany`, social icons...).

It adds a lot of confusion and can be way more simple if we separate
both widget.

Use CSS instead of JS
=====================
A JS function computed the size in pixel of the email to know if the
social icons should be on the second line or not. This is absurd because
it can be done in pure CSS.

Parse Odoo error messages
=========================
In some cases during the enrichment, the Odoo database might return an
error message instead of an "enrichment info".

Those messages were never displayed in the plugin.

Task-2601837